requirements.txt
# --- Manejo de datos y análisis
pandas
numpy

# --- Visualización
matplotlib
seaborn

# --- FastF1 para obtener datos de F1
fastf1

# --- Base de datos relacional (SQL Server)
pyodbc
sqlalchemy

# --- LLM local
requests
Ollama

# --- Telegram Bot
python-telegram-bot

# --- Logs, utilidades
python-dotenv